Checkpoint: steps 5-8 + deals export + Reeves + UX polish
Adds everything built on top of the initial step 1-4 commit:

* Per-well decline forecasting: /api/forecasts batch + preview + PATCH
  override, fit metrics, downtime detection, locked flag
  (steps 5-6: backend/app/forecasting/, backend/app/api/forecasts.py)
* Type curves: aggregate across wells, normalize per lateral foot,
  align by peak vs first-prod month, fit Arps per percentile, save +
  version + compare + manual tweak panel + CSV-zip export
  (backend/app/type_curves/, backend/app/api/type_curves.py,
   frontend/src/type_curves/, frontend/src/pages/TypeCurvePage.tsx)
* Auth + login flow (JWT bearer), users table, /api/auth/*
  (backend/app/api/auth.py, backend/app/core/security.py,
   backend/app/cli/create_user.py, frontend/src/pages/LoginPage.tsx)
* Map UX: PMTiles basemap, blocks + sections overlays, formation
  colors, lasso/box selection, well-detail modal, MVT tile pipeline
  (backend/app/wells_api/, frontend/src/components/MapView.tsx,
   frontend/src/panels/FilterPanel.tsx)
* Review tab: per-well grid with outlier flagging + EUR/ft + wellstick
  spatial view colored by EUR/ft
* Deals: group type curves and export the package as an Excel workbook
  (one metadata + one 50-yr fitted-forecast sheet per curve)
  (backend/app/api/deals.py, backend/app/db/models/deals.py,
   migration 0007_deals, frontend/src/api/deals.ts)
* Multi-county scope: Loving + Reeves as the canonical default; CLI
  and /api/sync/run accept comma-separated counties; synthetic client
  uses per-county FIPS + bbox tables
* Map + Review hover tooltips with well info; Review-tab color-ramp
  legend; ForecastDetailModal cum-chart continuity fix so the cum
  forecast picks up from where cum actual leaves off
* Infra: Postgres backup scripts, OTLS → blocks/sections shapefile
  conversion scripts, alembic revisions 0002-0007

+++ backend/app/db/models/production_monthly.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import date
 
-from sqlalchemy import Date, Float, ForeignKey, Integer, String
+from sqlalchemy import Date, Float, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -25,7 +25,10 @@ class ProductionMonthly(Base):
     oil_bbl: Mapped[float | None] = mapped_column(Float)
     gas_mcf: Mapped[float | None] = mapped_column(Float)
     water_bbl: Mapped[float | None] = mapped_column(Float)
-    producing_days: Mapped[int | None] = mapped_column(Integer)
+    # Float because Enverus reports fractional producing-days for partial
+    # first months (e.g. 0.166 = ~4 hours online). Integer truncation
+    # would zero-out IP-rich first-month rates.
+    producing_days: Mapped[float | None] = mapped_column(Float)
 
     # Diagnostic only — DO NOT use in forecasting or type curve aggregation.
     rate_prodday_bopd: Mapped[float | None] = mapped_column(Float)
